Add sequelize helpers :tada:
- utilizing helpers inside server scripts
